apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1
boto3==1.34.26
redis==5.0.1
arq==0.25.0
httpx==0.26.0
python-dotenv==1.0.0
